Moved query types
